Currently the only way to induce a bug is start rs for an ssl_id which which does not have a password saved, then try use an ssl id at start dialog  which does or does not have a pword (help_pword:bool enters an illegal state)
I have fixed it by checking if rsinit::passwd == "" then trying to load for that ssl_id,
if that fails then i set help_pword to false; does not have password and pass flow to gpg to ask for passwd. (problem was rs thought user had help/pass file and they didn't).

i have checked it in and tested for
1. starting with ssl_id which has saved pword, and switching ids at start dialog
2. secondary id may or may not have saved pword. 


git-svn-id: https://retroshare.svn.sourceforge.net/svnroot/retroshare/trunk@2801 56268a12-702b-0410-a8f4-c53eb35c22f6